AVS-421_67 - M2.2 Support
- Refactor code to remove abstract class for conditionally loading new
  class in constructor
